+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `HttpCaching` concern in `ApiResource`, which can make it easier to set Http Cache Control headers.
 * `HasArbitraryData` interface and a default implementation in `ArbitraryData` concern, as part of the utils package. 
 * `sync()` method added for `FileStream`. [#105](https://github.com/aedart/athenaeum/issues/105)
+* Test `Response` utility.
 
 ### Changed